Rename networkcontroller to basecontroller
