+++ controller/__init__.py
@@ -5,6 +5,6 @@
 
 This module includes all classes which relate to the controller portion of the MVC pattern.
 '''
-from .networkcontroller import NetworkController
+from .basecontroller import InputController
 from .mainwindowcontroller import MainWindowController
 from .activationscontroller import ActivationsController
